refactor: 컴포넌트 page 폴더로 이동
