+++ src/auth/HomePrivate.tsx
@@ -2,8 +2,8 @@ import React, { Suspense } from 'react';
 import { Outlet, Navigate } from 'react-router-dom';
 import { useSelector } from 'react-redux';
 import { RootState } from 'states/store';
-import AdminHomePage from 'components/admin-MainPage/AdminHomePage';
-import AdminNoGroupPage from 'components/admin-MainPage/AdminNoGroupPage';
+import AdminHomePage from 'pages/admin-MainPage/AdminHomePage';
+import AdminNoGroupPage from 'pages/admin-MainPage/AdminNoGroupPage';
 import OnBoardingPage from 'pages/OnBoardingPage';
 
 const HomePrivate = () => {
