+++ src/components/@commons-feature/calendar/Calendar.tsx
@@ -5,7 +5,7 @@ import Text from 'components/@commons/Text';
 import { Badge, BadgeCont, DateCont, InnerDayBox, MonthBox, OutterDayBox, WeekGrid } from './CalendarStyle';
 import FlexContainer from '../../@commons/FlexContainer';
 import { NextButton, PrevButton } from '../../@commons/iconButtons';
-import { dateAtom } from 'components/admin-MainPage/AdminScheduleSection';
+import { dateAtom } from 'pages/admin-MainPage/AdminScheduleSection';
 import { useAtom } from 'jotai';
 
 interface Props {
